# swagger/template.yml
swagger: "2.0"
info:
  title: edu_forms API
  description: Multi-tenant educational assessment platform
  version: "1.0.0"
host: "127.0.0.1:5000"
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT access token. Example: Bearer eyJhbG..."
  WorkspaceId:
    type: apiKey
    name: X-Workspace-Id
    in: header
    description: Active workspace context (not in JWT)
...
paths:
...
  /reports:
    post:
      tags: [Reports]
      summary: Create a support report
      description: |
        Authenticated users can submit a report/support request to platform administration.
        The backend derives `created_by_user_id` from the JWT and optionally derives
        `workspace_id` from the `X-Workspace-Id` header. The request body must not include
        `user_id`, `workspace_id`, or `status`.
      security: [{ Bearer: [] }]
      parameters:
        - in: header
          name: X-Workspace-Id
          type: integer
          required: false
          description: Optional active workspace context for the report
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [title, description, category]
            properties:
              title:
                type: string
                example: "Cannot access my exam"
              description:
                type: string
                example: "The exam page keeps loading and never starts."
              category:
                type: string
                enum: [TECHNICAL, ACCOUNT, CONTENT, CHEATING, OTHER]
                example: TECHNICAL
      responses:
        201:
          description: Report created
          schema:
            type: object
            properties:
              message: { type: string, example: "Report created successfully" }
              report_id: { type: integer, example: 1 }
        401:
          description: Missing or invalid token
        403:
          description: User account is not active or workspace membership is invalid
        422:
          description: Validation error
    get:
      tags: [Reports]
      summary: List my reports
      description: Returns only reports created by the authenticated user.
      security: [{ Bearer: [] }]
      responses:
        200:
          description: User reports
          schema:
            type: object
            properties:
              reports:
                type: array
                items:
                  type: object
                  properties:
                    report_id: { type: integer, example: 1 }
                    title: { type: string, example: "Cannot access my exam" }
                    category: { type: string, example: TECHNICAL }
                    status: { type: string, example: UNREAD }
                    created_at: { type: string, format: date-time }
        401:
          description: Missing or invalid token
  /reports/{report_id}:
    get:
      tags: [Reports]
      summary: Get my report details
      description: Returns details for a report created by the authenticated user.
      security: [{ Bearer: [] }]
      parameters:
        - name: report_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Report details
          schema:
            type: object
            properties:
              report_id: { type: integer, example: 1 }
              title: { type: string, example: "Cannot access my exam" }
              description: { type: string, example: "The exam page keeps loading and never starts." }
              category: { type: string, example: TECHNICAL }
              status: { type: string, example: UNREAD }
              created_at: { type: string, format: date-time }
              updated_at: { type: string, format: date-time }
        401:
          description: Missing or invalid token
        403:
          description: Report does not belong to user
        404:
          description: Report not found
...
  /api/super-admin/reports:
    get:
      tags: [Super Admin Management]
      summary: List all support reports
      description: |
        Platform-wide support report listing for Super Admins.
        Supports pagination and optional filtering by status and category.
      security: [{ Bearer: [] }]
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: per_page
          in: query
          type: integer
          default: 20
        - name: status
          in: query
          type: string
          enum: [UNREAD, IN_REVIEW, REVIEWED, REJECTED]
        - name: category
          in: query
          type: string
          enum: [TECHNICAL, ACCOUNT, CONTENT, CHEATING, OTHER]
      responses:
        200:
          description: Paginated report list
          schema:
            type: object
            properties:
              reports:
                type: array
                items:
                  type: object
                  properties:
                    report_id: { type: integer, example: 1 }
                    title: { type: string, example: "Cannot access my exam" }
                    category: { type: string, example: TECHNICAL }
                    status: { type: string, example: UNREAD }
                    created_at: { type: string, format: date-time }
                    reporter:
                      type: object
                      properties:
                        user_id: { type: integer, example: 10 }
                        full_name: { type: string, example: "Ahmed Ali" }
                        email: { type: string, example: "student@example.com" }
              total: { type: integer, example: 50 }
              page: { type: integer, example: 1 }
              per_page: { type: integer, example: 20 }
              pages: { type: integer, example: 3 }
        403:
          description: Super admin access required
  /api/super-admin/reports/{report_id}:
    get:
      tags: [Super Admin Management]
      summary: Get support report details
      description: |
        Returns full report details for Super Admins. If the report is UNREAD,
        opening it changes its status to IN_REVIEW.
      security: [{ Bearer: [] }]
      parameters:
        - name: report_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Report details
          schema:
            type: object
            properties:
              report_id: { type: integer, example: 1 }
              title: { type: string, example: "Cannot access my exam" }
              description: { type: string, example: "The exam page keeps loading and never starts." }
              category: { type: string, example: TECHNICAL }
              status: { type: string, example: IN_REVIEW }
              created_at: { type: string, format: date-time }
              updated_at: { type: string, format: date-time }
              reporter:
                type: object
                properties:
                  user_id: { type: integer, example: 10 }
                  full_name: { type: string, example: "Ahmed Ali" }
                  email: { type: string, example: "student@example.com" }
                  role: { type: string, example: STUDENT }
                  workspace_name: { type: string, example: "ABC Academy" }
        403:
          description: Super admin access required
        404:
          description: Report not found
    patch:
      tags: [Super Admin Management]
      summary: Update support report status
      description: |
        Updates only the report status. Title, description, reporter, and workspace
        are immutable through this endpoint.
      security: [{ Bearer: [] }]
      parameters:
        - name: report_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [status]
            properties:
              status:
                type: string
                enum: [UNREAD, IN_REVIEW, REVIEWED, REJECTED]
                example: IN_REVIEW
      responses:
        200:
          description: Report status updated
          schema:
            type: object
            properties:
              message: { type: string, example: "Report status updated successfully" }
              report:
                type: object
                properties:
                  report_id: { type: integer, example: 1 }
                  status: { type: string, example: IN_REVIEW }
        403:
          description: Super admin access required
        404:
          description: Report not found
        422:
          description: Invalid report status
  /api/super-admin/organizations/{organization_id}/suspend:
    patch:
      tags: [Super Admin Management]
      summary: Suspend an organization
      description: |
        Suspends a workspace organization (SOLO or INSTITUTION). Data and memberships are preserved.
        Sets workspace status to SUSPENDED and records suspension metadata.
        Suspended organizations are excluded from active workspace access and login workspace selection.
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: organization_id
          in: path
          type: integer
          required: true
          description: Workspace id (`workspaces.kind` is SOLO or INSTITUTION)
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [reason]
            properties:
              reason:
                type: string
                example: "Violation of platform policies"
      responses:
        200:
          description: Organization suspended
          schema:
            type: object
            properties:
              message: { type: string, example: "Organization suspended successfully" }
              organization:
                type: object
                properties:
                  id: { type: integer, example: 1 }
                  name: { type: string, example: "ABC Academy" }
                  organization_type: { type: string, example: INSTITUTION }
                  status: { type: string, example: SUSPENDED }
                  suspended_at: { type: string, format: date-time }
                  suspension_reason: { type: string, example: "Violation of platform policies" }
                  students_count: { type: integer, example: 500 }
                  teachers_count: { type: integer, example: 35 }
                  tests_count: { type: integer, example: 120 }
        404:
          description: Organization not found
        409:
          description: Organization is already suspended
        422:
          description: Validation error
        403:
          description: Super admin access required
  /api/super-admin/organizations/{organization_id}/restore:
    post:
      tags: [Super Admin Management]
      summary: Restore a suspended organization
      description: |
        Restores a previously suspended workspace organization (SOLO or INSTITUTION).
        Sets status to ACTIVE and clears suspension metadata.
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: organization_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Organization restored
          schema:
            type: object
            properties:
              message: { type: string, example: "Organization restored successfully" }
              organization:
                type: object
                properties:
                  id: { type: integer, example: 1 }
                  name: { type: string, example: "ABC Academy" }
                  organization_type: { type: string, example: INSTITUTION }
                  status: { type: string, example: ACTIVE }
                  students_count: { type: integer, example: 500 }
                  teachers_count: { type: integer, example: 35 }
                  admins_count: { type: integer, example: 3 }
                  tests_count: { type: integer, example: 120 }
                  attempts_count: { type: integer, example: 5400 }
        404:
          description: Organization not found
        422:
          description: Organization is not suspended
        403:
          description: Super admin access required
  /api/super-admin/organizations/{organization_id}:
    get:
      tags: [Super Admin Management]
      summary: Get organization details
      description: |
        Complete management view for a specific workspace organization (SOLO or INSTITUTION).
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: organization_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Organization details
          schema:
            type: object
            properties:
              id: { type: integer, example: 1 }
              name: { type: string, example: "ABC Academy" }
              slug: { type: string, example: "abc-academy" }
              organization_type: { type: string, example: INSTITUTION }
              status: { type: string, example: ACTIVE }
              created_at: { type: string, format: date-time }
              updated_at: { type: string, format: date-time }
              suspended_at: { type: string, format: date-time }
              suspension_reason: { type: string }
              is_verified_by_superadmin: { type: boolean, example: true }
              owner:
                type: object
                properties:
                  user_id: { type: integer, example: 5 }
                  name: { type: string, example: "Owner Name" }
                  email: { type: string, example: "owner@abc.edu" }
              admins_count: { type: integer, example: 3 }
              students_count: { type: integer, example: 500 }
              teachers_count: { type: integer, example: 35 }
              active_users_count: { type: integer, example: 538 }
              tests_count: { type: integer, example: 120 }
              attempts_count: { type: integer, example: 5400 }
        404:
          description: Organization not found
        403:
          description: Super admin access required
  /api/super-admin/users/{user_id}/suspend:
    patch:
      tags: [Super Admin Management]
      summary: Suspend a user account
      description: |
        Suspends a user account without deleting data or memberships.
        Sets user status to SUSPENDED, records suspension metadata, and revokes active sessions.
        Super admins cannot suspend their own account.
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: user_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [reason]
            properties:
              reason:
                type: string
                example: "Repeated policy violations"
      responses:
        200:
          description: User suspended
          schema:
            type: object
            properties:
              message: { type: string, example: "User suspended successfully" }
              user:
                type: object
                properties:
                  id: { type: integer, example: 15 }
                  name: { type: string, example: "Ahmed Ali" }
                  email: { type: string, example: "user@example.com" }
                  status: { type: string, example: SUSPENDED }
                  suspended_at: { type: string, format: date-time }
                  suspension_reason: { type: string, example: "Repeated policy violations" }
                  roles:
                    type: array
                    items: { type: string }
                    example: [TEACHER]
                  organizations:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer, example: 2 }
                        name: { type: string, example: "ABC Academy" }
                        organization_type: { type: string, example: INSTITUTION }
                        status: { type: string, example: ACTIVE }
        403:
          description: Forbidden (e.g. self-suspension)
        404:
          description: User not found
        409:
          description: User is already suspended
        422:
          description: Validation error
  /api/super-admin/users/{user_id}/restore:
    post:
      tags: [Super Admin Management]
      summary: Restore a suspended user
      description: |
        Restores a previously suspended user account.
        Sets status to ACTIVE and clears suspension metadata.
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: user_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: User restored
          schema:
            type: object
            properties:
              message: { type: string, example: "User restored successfully" }
              user:
                type: object
                properties:
                  id: { type: integer, example: 15 }
                  name: { type: string, example: "Ahmed Ali" }
                  email: { type: string, example: "user@example.com" }
                  status: { type: string, example: ACTIVE }
                  roles:
                    type: array
                    items: { type: string }
                    example: [TEACHER]
                  organizations:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer, example: 2 }
                        name: { type: string, example: "ABC Academy" }
                        organization_type: { type: string, example: INSTITUTION }
                        status: { type: string, example: ACTIVE }
        404:
          description: User not found
        422:
          description: User is not suspended
        403:
          description: Super admin access required
  /api/super-admin/users/{user_id}:
    get:
      tags: [Super Admin Management]
      summary: Get user details
      description: |
        Complete management view for a specific user, including memberships and activity stats.
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: user_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: User details
          schema:
            type: object
            properties:
              id: { type: integer, example: 15 }
              name: { type: string, example: "Ahmed Ali" }
              email: { type: string, example: "user@example.com" }
              status: { type: string, example: ACTIVE }
              created_at: { type: string, format: date-time }
              updated_at: { type: string, format: date-time }
              last_login_at: { type: string, format: date-time }
              last_activity_at: { type: string, format: date-time }
              email_verified: { type: boolean, example: true }
              is_superadmin: { type: boolean, example: false }
              suspended_at: { type: string, format: date-time }
              suspension_reason: { type: string }
              roles:
                type: array
                items: { type: string }
                example: [TEACHER]
              organizations:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer, example: 2 }
                    name: { type: string, example: "ABC Academy" }
                    organization_type: { type: string, example: INSTITUTION }
                    status: { type: string, example: ACTIVE }
              memberships:
                type: array
                items:
                  type: object
                  properties:
                    membership_id: { type: integer, example: 42 }
                    role: { type: string, example: TEACHER }
                    status: { type: string, example: ACTIVE }
                    joined_at: { type: string, format: date-time }
              created_tests_count: { type: integer, example: 12 }
              completed_attempts_count: { type: integer, example: 48 }
        404:
          description: User not found
        403:
          description: Super admin access required
  /api/super-admin/users:
    get:
      tags: [Super Admin Management]
      summary: List all platform users with memberships and roles
      description: |
        Platform-wide user listing for Super Admin user management.
        Supports pagination and filtering by role, organization, status, and search.
        Super Admin only.
      security: [{ Bearer: [] }]
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: per_page
          in: query
          type: integer
          default: 20
        - name: role
          in: query
          type: string
          enum: [ADMIN, TEACHER, STUDENT, SUPER_ADMIN]
        - name: organization_id
          in: query
          type: integer
          description: Filter users by organization workspace id
        - name: status
          in: query
          type: string
          enum: [PENDING_VERIFICATION, PENDING_APPROVAL, REGISTRATION_REJECTED, ACTIVE, SUSPENDED, DISABLED]
        - name: search
          in: query
          type: string
          description: Search by user name or email
      responses:
        200:
          description: Paginated user list
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer, example: 15 }
                    name: { type: string, example: "Ahmed Ali" }
                    email: { type: string, example: "user@example.com" }
                    status: { type: string, example: ACTIVE }
                    created_at: { type: string, format: date-time }
                    suspended_at: { type: string, format: date-time }
                    suspension_reason: { type: string }
                    roles:
                      type: array
                      items: { type: string }
                      example: [TEACHER]
                    organizations:
                      type: array
                      items:
                        type: object
                        properties:
                          id: { type: integer, example: 2 }
                          name: { type: string, example: "ABC Academy" }
                          organization_type: { type: string, example: INSTITUTION }
                          status: { type: string, example: ACTIVE }
              total: { type: integer, example: 120 }
              page: { type: integer, example: 1 }
              per_page: { type: integer, example: 20 }
              pages: { type: integer, example: 6 }
        403:
          description: Super admin access required
...